Implement `KartCtrl::CheckChange` function.
This code seems really strange... not only do the offsets not correlate to proper field members of the KartBody class itself, (specifically `mDriverModels` being index 2 and 3, not 0 and 1, meaning its out of bounds of the array), but the structure of the jumps don't make sense either...

It's like someone used explicit gotos, or the compiler optimised something in a weird way when it comes to values returned.

Yuck, yuck, yuck! - Work out nicer C code later.