Implment `KartCtrl` functions `IsWallReact` and `HaveBalloon`
`IsWallReact` relies on `KartTumble` which had the array `_0` listed as private. As such, I've changed this to public for now. We can tweak permissions at a later point if needed. Additionally, the return type for IsWallReact looks like it should be a `bool`.

`HaveBalloon` is interesting, as it relies on a compiler macro `__cntlzw`, which counts the number of leading zeros of a 32-bit value.

Also stage missed function prototype for `KartCtrl::IsBurn`.