Implement `KartCtrl::GetPortPtr` function.
Note that the return for this function probably has the wrong type, as the way `viewdata` is accessed is a bit obscure.

Also, this great example of how nested switch/case statements are used. It was a pain to decompile, but a great way to see how the nested structures in C compile to assembly.