Very rough implementation of the inline functions for `GeographyObjManager::getKartReactHitObjectList` and `GeographyObjManager::getKartHistRefVecNormalObj`.
At this point I haven't been able to work out enough about the GeographyObjManager class to work out the appropriate types for each class property, so instead we get the base address, calculate the offset and then cast the memory location to a pointer to the object we want the data to be represented as.

This will need to be changed down the line, but for now it allows `KartDossin` to be fully decompiled.