Decompilation of `KartDossin`.
The functions here handle the effect when a player's kart gets flattened like a pancake by a course hazard, such as Dino Dino's feet in Dino Dino Jungle, or the Thwomps from Bowser's Castle.

There may be more instances where the effect is used, but these two are the easiest I've found to test.

Also, thanks to mousepascals (@muff1n1634 on Discord) for help with the `DoAfterDossinCrl` decompilation - sometimes you just need a second pair of eyes on your code. :-)